requirements.txt
babel==2.17.0
black==25.1.0
chardet==5.2.0
click==8.1.8
contourpy==1.3.1
coverage==7.6.12
cycler==0.12.1
et_xmlfile==2.0.0
flake8==7.1.2
fonttools==4.56.0
iniconfig==2.0.0
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pyparsing==3.2.1
pytest==8.3.5
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
reportlab==4.3.1
six==1.17.0
tkcalendar==1.6.1
tzdata==2025.1
XlsxWriter==3.2.2
